Auto Align

This function will automatically align your data based on the smoothing method you choose.

Usage

auto_align(
  df1,
  df2,
  rt_lower = -0.5,
  rt_upper = 0.5,
  mz_lower = -15,
  mz_upper = 15,
  rt_smooth = 0.2,
  mz_smooth = 0.2,
  minimum_intensity = 1000,
  rt_simplification = 0.5,
  mz_simplification = 15,
  smooth_method = "lowess",
  multipliers = c(6, 6, 6),
  weights = c(1, 1, 1),
  keep_features = c(F, F)
)

Arguments

df1

A data frame representing the results of a preprocessed LC-MS experiment.

df2

A data frame of representing the results of a second preprocessed LC-MS experiment.

rt_lower

A numeric indicating the lower bound of the RT range to be considered for alignment.

rt_upper

A numeric indicating the upper bound of the RT range to be considered for alignment.

mz_lower

A numeric indicating the lower bound of the m/z range to be considered for alignment.

mz_upper

A numeric indicating the upper bound of the m/z range to be considered for alignment.

rt_smooth

A numeric indicating the smoothing parameter for RT.

mz_smooth

A numeric indicating the smoothing parameter for m/z.

minimum_intensity

A numeric indicating the minimum intensity to be considered for alignment.

rt_simplification

A numeric indicating the simplification parameter for RT.

mz_simplification

A numeric indicating the simplification parameter for m/z.

smooth_method

A character indicating the smoothing method to be used. Options are "lowess", "spline", and "gaussian".

multipliers

A numeric vector indicating the multipliers to be used for the alignment.

weights

A numeric vector indicating the weights to be used for the alignment.

keep_features

A logical vector indicating whether or not to keep features that are not matched.

Value

A data frame containing the aligned data.