Consolidate

Consolidate adducts/isomers with similar features into one observation

Usage

consolidate(
  ms_obj,
  use_rt = TRUE,
  use_mz = TRUE,
  rt_threshold = NULL,
  mz_threshold = NULL
)

Arguments

ms_obj

a massSight object

use_rt

boolean if retention time should be used in identifying metabolites that should be consolidated

use_mz

boolean if mass to charge ratio should be used in identifying metabolites that should be consolidated

rt_threshold

user defined retention time threshold for defining adducts/isomers

mz_threshold

user defined mass to charge ratio threshold for defining adducts/isomers

Value

a massSight object with consolidated metabolites