Create MS Object

Create an MSObject from a data frame.

Usage

create_ms_obj(
  df,
  name,
  id_name = "Compound_ID",
  rt_name = "RT",
  mz_name = "MZ",
  int_name = "Intensity",
  metab_name = "Metabolite"
)

Arguments

df

A data frame containing the raw data.

name

A character indicating the name of the experiment.

id_name

A character indicating the name of the column containing the compound IDs.

rt_name

A character indicating the name of the column containing the retention times.

mz_name

A character indicating the name of the column containing the m/z values.

int_name

A character indicating the name of the column containing the intensities

metab_name

An optional character indicating the name of the column containing the metabolite annotations

Value

An MSObject.